fix(notifications,settings): resolve override keys case-insensitively
- `overrides.ts`: both key forms (worktree identity and legacy checkout path) now compare case-insensitively, and a save folds away every case variant of both before writing under the freshly resolved casing, so no ghost entry survives that `overrideEntry` could return but no later save could reach.
- `NotificationsSection`: the own/other key partition and the recent-repo identity map key on the lowercased form, matching `overrideEntry`, so a differently-cased key no longer counts as both the open repo and another repository, or gets mislabeled "(not in recent repositories)".
- New `src/lib/notifications/matrix.ts` holds the shared vocabulary, the signature/count helpers, and the two open-state stores. Keeping it component-free lets the dialog module Fast-Refresh instead of full-invalidating, which showed up as ghost state at runtime.
- `MatrixTable` now requires a `caption`, so the global and per-repository grids are tellable apart under table navigation.
- Copy: README and the in-app guide describe the per-repo customize flow in current terms, and the `emit.ts` dedupe-key comment states what producer keys actually look like.
